+++ docs/14_0_0/gettingstarted/whatsnew.md
@@ -44,6 +44,9 @@ Look into [migration guide](https://primefaces.github.io/primefaces/14_0_0/#/../
     * Added `filterPlaceholder` for `Column` and `Columns`
     * Added `rowData` to `CellEditEvent` which contains the entire row from the cell being edited.
 
+* Dashboard
+    * Added `var` to allow dynamic panels in `DashboardWidget.setValue(obj)` per panel
+    
 * Messages
     * Added `clearMessages` widget method to clear all current messages.